+++ Services/Reserve/Reserve.WebApi/Core/Aggregates/RoomReserve.cs
@@ -2,7 +2,7 @@
 {
     public class RoomReserve
     {
-        public RoomReserve(Guid Id, HotelReserve Hotel, string RoomNumb, RoomReserveCapacity Capacity, RoomReserveStatus Status, long price)
+        public RoomReserve(Guid Id, HotelReserve Hotel, string RoomNumb, RoomReserveCapacity Capacity, RoomReserveStatus Status, double price)
         {
             this.Id = Id;
             this.Hotel = Hotel;
@@ -13,7 +13,7 @@
         }
         public Guid Id { get; private set; }
         public string RoomNumb { get; private set; }
-        public long Price { get; private set; }
+        public double Price { get; private set; }
         public HotelReserve Hotel { get; private set; }
         public RoomReserveCapacity Capacity { get; private set; }
         public RoomReserveStatus Status { get; private set; }
